Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main(){
     int a,b,c;
     float avg;
     scanf("%d%d%d",&a,&b,&c);
-    int sum=a+b+c;
+    float sum=a+b+c;
     avg=sum/3;
     printf("Average: %.2f",avg);
     return 0;
